requirements.txt
Pillow == 7.2.0
numpy == 1.19.1